Refactor sorting functions and utility methods to use long data type for improved range handling

+++ pu_radix_sort.c
@@ -48,8 +48,8 @@ static void simple_sort(t_stacks *stacks)
 
 void radix_sort(t_stacks *stacks)
 {
-    int min;
-    int bit;
+    long min;
+    long bit;
     int i;
     int size;
     long num;
